backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.0
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.2
langchain-community==0.0.10
composio-langchain==0.3.0

